requirements.txt
# Drowsy Driving Detection - Requirements

# Deep Learning Framework
torch>=1.12.0
torchvision>=0.13.0

# Numerical Computing
numpy>=1.21.0

# Image Processing
Pillow>=9.0.0

# Machine Learning Utils
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress Bar
tqdm>=4.64.0

# Optional: Data Analysis
pandas>=1.3.0

# Optional: Webcam Support
# opencv-python>=4.5.0

# Optional: Jupyter Notebook
# jupyter>=1.0.0
# ipywidgets>=7.6.0